refactor(fsmonitor): consolidate SIGTERM-poll-SIGKILL escalation
The two fsmonitor-daemon termination paths (`wt remove`'s
`force_kill_fsmonitor_via_socket` and `reap_orphan_fsmonitor_daemons`)
shipped near-identical SIGTERM-poll-SIGKILL implementations in adjacent
commits (#2813 and #2814). Lift `escalate_terminate`, `NixSignaller`,
and `REAP_KILL_DEADLINE` to `pub(crate)` and have `remove.rs` call them
in place of `terminate_pid`. Drops `FSMONITOR_SIGTERM_GRACE` in favor of
the single shared constant.

Move the real-process SIGTERM/SIGKILL coverage from `remove.rs` to
`fsmonitor.rs` so `NixSignaller` keeps its end-to-end signal-delivery
tests; the `FakeSignaller`-based tests already cover the escalation
logic.

Closes #2830.
